requirements.txt
-e git+https://github.com/UnicycleLabs/apilib@ffebb0851aa41b012bb9277544617ec902be65a2#egg=apilib
appier==1.7.39
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2016.9.26
cffi==1.9.1
click==6.6
cryptography==1.5.3
decorator==4.0.10
enum34==1.1.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
GeoAlchemy2==0.3.0
gnureadline==6.3.3
html5lib==1.0b8
idna==2.1
ipaddress==1.0.17
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
path.py==8.2.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.17
pyflakes==1.3.0
Pygments==2.1.3
pyOpenSSL==16.2.0
PyPDF2==1.26.0
python-dateutil==2.6.0
reportlab==3.3.0
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.3
traitlets==4.3.1
urllib3==1.19
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.11.11
xhtml2pdf==0.0.6
